+++ Sources/SuperwallKit/Paywall/Request/PaywallRequestManager.swift
@@ -40,11 +40,15 @@ actor PaywallRequestManager {
   ///  - Returns A paywall.
   func getPaywall(from request: PaywallRequest) async throws -> Paywall {
     let deviceInfo = factory.makeDeviceInfo()
+    let joinedSubstituteProductIds = request.overrides.products?.values
+      .sorted(by: { $0.productIdentifier < $1.productIdentifier })
+      .map { $0.productIdentifier }.joined()
+
     let requestHash = PaywallLogic.requestHash(
       identifier: request.responseIdentifiers.paywallId,
       event: request.eventData,
       locale: deviceInfo.locale,
-      paywallProducts: request.overrides.products
+      joinedSubstituteProductIds: joinedSubstituteProductIds
     )
 
     if var paywall = paywallsByHash[requestHash],